pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stock-analyzer"
version = "0.1.0"
description = "AI-based Stock Analysis System"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pandas>=2.0.0",
    "akshare>=1.14.0",
    "tushare>=1.2.0",
    "requests>=2.31.0",
    "openai>=1.0.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
analyze-stock = "stock_analyzer.main_pipeline:main"

[tool.setuptools]
packages = ["stock_analyzer", "stock_analyzer.api", "stock_analyzer.core", "stock_analyzer.utils"]
